Customize for Figma's use
The upstream maintainer is opinionated about what this project should and should not do in a way that diverges from what we need. That's their perogative! We'll maintain our own patches. It's not that complicated of a dependency.

We need:
* Support for `--experimental_convenience_symlinks=ignore`
* Not to require an `external` symlink inside the repository

See https://github.com/hedronvision/bazel-compile-commands-extractor/issues/14 and https://github.com/hedronvision/bazel-compile-commands-extractor/pull/75 for further discussion.